+++ R/zzz.R
@@ -3,6 +3,7 @@
   op <- options()
   op.styler <- list(
     styler.addins_style_transformer = "styler::tidyverse_style()",
+    styler.cache_root = NULL,
     styler.cache_name = styler_version,
     styler.colored_print.vertical = TRUE,
     styler.ignore_start = "# styler: off",
@@ -12,9 +13,32 @@
   )
   toset <- !(names(op.styler) %in% names(op))
   if (any(toset)) options(op.styler[toset])
+  ask_to_switch_to_non_default_cache_root()
   remove_cache_old_versions()
   remove_old_cache_files()
   invisible()
+}
+
+
+ask_to_switch_to_non_default_cache_root <- function(ask = interactive()) {
+  if (ask && runif(1) > 0.9 && is.null(getOption("styler.cache_root"))) {
+    cli::cli_inform(paste0(
+      "The R option `styler.cache_root` is not set, which means the cache ",
+      "will get cleaned up after 6 days (and repeated styling will be slower).",
+      " To keep cache files longer, set ",
+      "the option to location within the {{R.cache}} cache where you want to ",
+      "store the cache, e.g. `\"styler-perm\"`.\n\n",
+      "options(styler.cache_root = \"styler-perm\")\n\n",
+      "in your .Rprofile. Note that the cache literally ",
+      "takes zero space on your disk, only the inode, and you can always ",
+      "manually clean up with `styler::cache_clear()`, and if you update the ",
+      "{{styler}} package, the cache is removed in any case. To ignore this ",
+      "message in the future, set the default explictly to \"styler\" with\n\n",
+      "options(styler.cache_root = \"styler\")\n\nin your `.Rprofile`. This ",
+      "message will only be displayed once in a while.\n"
+    ))
+  }
+  options(styler.cache_root = "styler")
 }
 
 remove_old_cache_files <- function() {